+++ src/pages/ConsumerPurposeCreatePage/components/PurposeCreateEServiceAutocomplete.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'react-i18next'
 import { useFormContext } from 'react-hook-form'
 import { useAutocompleteTextInput } from '@pagopa/interop-fe-commons'
 import type { CatalogEService, CompactEService } from '@/api/api.generatedTypes'
-import type { PurposeCreateFormValues } from './PurposeCreateEServiceForm'
+import type { PurposeCreateFormValues } from './PurposeCreateForm'
 import { useQuery } from '@tanstack/react-query'
 import { DelegationQueries } from '@/api/delegation'
 import { AuthHooks } from '@/api/auth'
