+++ src/pages/ConsumerPurposeCreatePage/components/PurposeCreateTemplateAutocomplete.tsx
@@ -4,7 +4,7 @@ import { Alert } from '@mui/material'
 import React from 'react'
 import { useFormContext } from 'react-hook-form'
 import { useTranslation } from 'react-i18next'
-import type { PurposeCreateFormValues } from './PurposeCreateEServiceForm'
+import type { PurposeCreateFormValues } from './PurposeCreateForm'
 import { Spinner } from '@pagopa/interop-fe-commons'
 import { AuthHooks } from '@/api/auth'
 import { useQuery } from '@tanstack/react-query'
